+++ internal/serializers/ugorji/ugorji.go
@@ -1,6 +1,9 @@
-package goserbench
+package ugorji
 
-import "github.com/ugorji/go/codec"
+import (
+	"github.com/alecthomas/go_serialization_benchmarks/goserbench"
+	"github.com/ugorji/go/codec"
+)
 
 type UgorjiCodecSerializer struct {
 	codec.Handle
@@ -15,11 +18,11 @@ func (u *UgorjiCodecSerializer) Unmarshal(d []byte, o interface{}) error {
 	return codec.NewDecoderBytes(d, u.Handle).Decode(o)
 }
 
-func NewUgorjiCodecMsgPack() Serializer {
+func NewUgorjiCodecMsgPack() goserbench.Serializer {
 	return &UgorjiCodecSerializer{&codec.MsgpackHandle{}}
 }
 
-func NewUgorjiCodecBinc() Serializer {
+func NewUgorjiCodecBinc() goserbench.Serializer {
 	h := &codec.BincHandle{}
 	h.AsSymbols = 0
 	return &UgorjiCodecSerializer{h}
